[REFACTOR] Calculator 내부의 enum 클래스 분리
